File: 그룹 상세 데이터 조회 groupDetail.ts -> group.ts로 이동
